feat: adapt CANN of onnxruntime (PR #428)
