+++ python/rapidocr/inference_engine/onnxruntime/__init__.py
@@ -1,0 +1,4 @@
+# -*- encoding: utf-8 -*-
+# @Author: SWHL
+# @Contact: liekkaskono@163.com
+from .main import OrtInferSession
